Touch Input, Player Movement and rotated to face the right-side, also cannot leave the borders,
c. The Player’s Ship starts on the Left-hand side of the screen and in a fixed X-position (similar to having a fixed Y-position at the bottom of the screen). (5 points)
d. The Player’s Ship will move with Touch Input. (5 points)
e. The Player's Ship will be rotated to face the right-side of the screen. (5 points)
g. The Player’s Ship’s motion will be limited so that it cannot leave the borders of the screen. (5 points)