Enemy position, move from the top to the bottom.
j. The Enemy Game Objects will be positioned at the right-side of the screen. (5 points)
k. The Enemy Game Objects will move from the top to the bottom of the screen instead of moving left and right. They will reverse direction when they hit the borders of the screen. (5 points)